Stock price download
Downloads stock price data for stocks held by the funds at the date of each filing realease. The script extracts the stock tickers from the holdings CSV. Downloads data in batches of 5 with a sleep time to avoid a too many requests error that an earlier version of the script produced. Outputs a csv that includes the price changes of the stocks in question.