Correlation new
New working version of the script. Pulls Price Data for MSTR and Bitcoin and calculates the correlation coefficients for the time in between the earnings dates. The next step would be to also import the balance sheet of MSTR and calculate the bitcoin percentage of the total assets for further analysis. The script only starts the data in 2023 although both asset prices have longer history. Maybe try and find out why that is tmrw.